Make the command move files progressively.
The previous version used to find all the files to move and then do
the moves which requires a lot of memory for large amounts of files.
We now move files as they are found and only keep a map of which moved
file was where before.

Also per-used directory are now owned by the respective user and set
to mode 0700 for privacy.